# swagger.yaml
openapi: 3.0.0
info:
  title: LMS API
  description: Learning Management System API
  version: 1.0.0
servers:
  - url: 'http://localhost:5000/api'
paths:
  /users:
    get:
      summary: Get all users
      description: Returns a list of users in the LMS system.
      responses:
        '200':
          description: A list of users
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    username:
                          string
                    firstname:
                      type: string
                    lastname:
                      type: string
                    email:
                      type: string
                    role:
                      type: string
    post:
      summary: Add a new user
      description: Creates a new user in the LMS system with the provided details.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - firstname
                - lastname
                - email
                - password
                - role
              properties:
                username:
                  type: string
                  description: The username of the user.
                firstname:
                  type: string
                  description: The first name of the user.
                lastname:
                  type: string
                  description: The last name of the user.
                email:
                  type: string
                  description: The email address of the user.
                password:
                    type: string
                    description: The user's password.
                    pattern: '^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[+%&()!])[A-Za-z\d+%&()!]{8,}$'
                    minLength: 8
                    example: 'Abc123!+'
                role:
                  type: string
                  description: Role of the user (student or teacher).
                  enum:
                    - student
                    - teacher
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  firstname:
                    type: string
                  lastname:
                    type: string
                  email:
                    type: string
                  role:
                    type: string
        '400':
          description: Invalid input
  /users/{id}:
    get:
      summary: Get a single user
      description: Returns a single user from the LMS system.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: A single user
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  firstname:
                    type: string
                  lastname:
                    type: string
                  email:
                    type: string
                  role:
                    type: string
        '404':
          description: User not found